Parse multiple messages when available
This change reworks the parsing API quite a bit. I am happy with
this current implementation; the API enforces streaming of large
read/write messages, requires very little ongoing allocation (only
unavoidable cost is that of the io.ReadClosers for Twrite/Rread reqs),
and should allow for very fast decoding (it doesn't really decode; it
just validates all the sizes/offsets are OK and parses the bytes on
demand, like FlatBuffers).

The biggest change to the API here is that `Scanner.Msg`, which
returned the first message parsed in `Next`, is replaced with
`Scanner.Msgs`, which returns the last N messages read in the
input. ReadMsg will parse multiple messages from a single
Read call, if the data are available.

The reason I wanted to make this change was that there are certain
scenarios where it is legal to handle messages concurrently (throwing
order out the window). For instance, if the messages are coming from
different sessions on the same connection, it should be legal to handle
these messages concurrently in most cases. If we only had access to
one message at a time, all clients on the same connection would have
to wait for each message to be handled.

By returning a slice of messages, we allow callers to handle multiple
messages concurrently, so long as they finish with all of them
before fetching the next batch of messages.

In change to the API, invalid messages are no longer considered errors,
but are their own values (of type `BadMessage`), and do not stop
scanning. This makes a clearer distinction between parse errors and
connection errors.