Write messages to an io.Writer instead of a []byte
When I first started writing the "styx" package, I realized very
quickly that it was more important to be able to write messages
to a connection, and maintaining an intermediary []byte buffer
was wasteful and overlapped with the functionality of the bufio
package. This commit swaps out the NewT* and NewR* functions,
which created a 9P message in a provided byte slice, for Write*
functions that instead write it to an io.Writer.