Change WriteTwrite/WriteRread to take a byte slice.
When I first wrote these two functions, I was obsessed with minimizing
copying and memory usage. After trying to use these functions writing
the higher-level server implementation, I realize there are downsides
to accepting an io.Reader instead of a byte slice.

- You have to manage a separate count that may or may not be accurate.
- The io.Reader can return an error that must be checked

When I first wrote the functions to create messages, they wrote to a
byte slice and were held in memory. There it made sense to keep the
data portion, which could be up to 4GB, out of memory. With commit
c42f0c4, Twrite and Rread messages are written straight to the wire.
Taking a `[]byte` reduces book keeping and more closely mirrors the
types in the `io` package.