Add a TxWriter utility type.
For the multi-threaded server implementation, multiple threads may
want to write a message at any given time. The TxWriter utility type
provides a way to isolate sets of writes so that we don't mangle the
data stream.

It is common to see types that put a mutex around the Write method
to achieve a similar goal. That was not enough in my case, because
the functions that create messages make many Write calls. Arranging
to have the whole message in a single Write call is more work than
implementing these transactions.