Working on server main loop
This is still not functional, but checking this in anyway for
posterity. When a `*styx.Server` is serving a client, we have the following
stack of resources:

- A `*styx.session` (sessions are multiplexed over a connection),
  which sits on top of
- A `*styx.Conn`, one per connection, which contains
- A `*styxproto.Decoder`, that decodes incoming messages from
- A `*bufio.Reader`, which buffers for
- A `net.Conn`

The `*styx.Conn` also holds a `*bufio.Writer` to send messages back
to the client. This layout was heavily inspired by the `net/http` package.
I like the way things are progressing, although I am slightly concerned
with the growing complexity.