Add styxproto.Encoder type
An `Encoder` mirrors a `Decoder` and is used to produce messages.
This replaces the Write[Msg] functions, and has the following benefits:

- Fewer arguments to pass to the encoding function (stored `io.Writer`)
- Less chance of writing to the wrong `io.Writer`
- Can hide complexity around isolating writes from higher-level code
- Less typing: `enc.Rversion(msize, "9P2000")` vs `styxproto.Rversion(w, msize, "9P200")`

My biggest issue with the way things were before was how exposed the
`wire.TxWriter` handling was. It was littered throughout the `styx` package.
With this commit, all the lock and transaction management is hidden
behind the `Encoder` type, and there's no reason for it to leak out yet.