Add a callback-based server to the styxproto package.
I've been struggling with the implementation of the server portion
of the `styx` package; how to design the API, the implementation,
supporting stuff like cancellation, etc. I was considering too-many
things at once. So I chose to start by making an interface,
`styxproto.Server`, that provided one method for each type of transaction.
The `styxproto.Serve` function serves a single connection, and handles
the following:

- Message parsing
- Protocol version negotiation
- Monitoring and cancellation of pending transactions

Soon to come will be a dumb `styxproto.Server` implementation
that just logs its arguments, both for testing and as an example.