Facter out styxproto.Serve into styxserver package
The `styxproto.Serve` function and its related type are
markedly different from the rest of the `styxproto` package,
in that they are not focused on simply parsing or producing
9P messages, and instead focus on actually running a 9P
server.

Seeing as this functionality did not use any unexported features
of the `styxproto` package, it made sense to me to move it into
its own package.