Remove styxserver package
It was unclear what the `styxserver` package was for. I developed
it because I didn't know what direction I should go for the client/server
API in the `styx` package, and tried to mimic the `Srv` type in the plan 9
9P library. In the interest of keeping this package focused, I am removing
the `styxserver` package. I may resurrect parts of it should they be necessary
in the future.