Semi-functional server interface
With this commit I can get a rudimentary test case to run.
I am beginning to doubt my approach; I feel that the implementation
is beginning to bend under its own weight. I want to preserve and
expose the statefulness of 9P with the Serve9P interface, but it is
proving to be difficult to support. I am going to go and study some
other implementations for inspiration.