Add helper type for generating Stat structures from Readdir()
It's a bother to marshal os.FileInfo into a Stat structure yourself,
and then you have to package it up into a type that supports
ReadAt. The `*dirReader` type (accessed through the `styxfile.NewDir`
function) does this for you for any type that implements the method

	Readdir(n int) ([]os.FileInfo, error)

`*os.File` implements this, and many lookalikes do as well.