Don't close the connection on a malformed (but recoverable) message
I go back and forth on this one; it is certainly simpler to be strict
about message validation and shut people out at the first sign of
a bad message. However, I have been testing the implementation
interactively (via tools such as `9pfuse`) and have found it unpleasant
to be shut out due to certain verification routines being too strict.

I will definitely look for guidance on the right thing to do here.