The end of a session is not the end of its connection
Some clients, such as `9p`, create and cleanup a new
session for every command they run. Before this change,
the first call to, say, `9p ls`, would result in the connection
being closed after the last fid was clunked. It is perfectly
valid for a client to hold a connection open and Tattach
to create a new session "on-demand".