Fix inconsistent parsing of Stat structure
A few commits back I fixed parsing of Twstat messages
that forgot about the fid argument, but failed to update
the corresponding accessor method. This commit fixes
that, and in addition changes styxproto to always include
the leading size[2] in the Stat structure, as NewStat does
(but `Twstat.Stat()` and `Rstat.Stat()` did not).