Clear up misunderstanding about Twstat, Rstat messages
The `Twstat` and `Rstat` messages look like this:

	size[4] Rstat tag[2] stat[n]
	size[4] Twstat tag[2] fid[4] stat[n]

And a `stat` structure starts with the following field

	size[2]
		total byte count of the following data

My misunderstanding was thinking that the 2-byte `[n]` that preceded the
`stat` structure and the `size` field of the `stat` structure were the same
field, but that is not the case (and I am surprised that I read the documentation
in such a way); the size is copied (minus 2). So the message looks like

	size[4] Twstat tag[2] fid[4] n[2] size[2] type[2] dev[4] ...

Where `size` is (usually) `n-2`.