Slim down server API and remove bare channels
I copied a lot of ideas from the `net/http` package, like the `ServeMux`
for routing requests. While I still think request routing is a good idea,
I want its implementation to be informed by a real example server. The
Serve9P API is a little different from the ServeHTTP API -- a Serve9P
function lasts for the duration of an entire session, not a single request.
I want to re-think the API for routing requests.

While I wanted to expose a Request channel in my API, in order to easily
allow for things like timeouts in a `select{}` statement, I ultimately ended
up making the `Session.Requests` channel private, hidden behind a `Next`
method. The reason I did this was for cleanup; I wanted to allow users to
be able to ignore message types they did not care about, and have the
styx package make a "default" response for those messages.

We can assume a Serve9P function is "done" with a message when it
reads the next one from the channel. This is all and well, until we introduce
the concept of middleware. What if a Serve9P function wants to feed a message
to another Handler? How will it know when that Handler is done with the message?
It cannot try sending a new message because it doesn't have one; as soon as it
reads another message from its own `Requests` channel, the styx package would
be free to clean up the previous request that the downstream Handler may still be
processing. No good.

Using a function call to signal that a handler is done with a message gives us more
flexibility in signalling that the resources associated with that request can be cleaned
up.