Fix bug in detecting unhandled request
The styx package sends full types to sessions, not pointers.
And all methods on those requests operate on the type, not
a pointer to them. So setting the 'sent' member after a request
was handled did nothing. This change moves this signalling
outside of the request types and into the session/conn structs.
For the fake requests, we use a lookup table within the request
assemblers.