Add support for combining many handlers into one
This commit introduces `styx.Stack`, which runs each request through a
series of handlers until one responds. This provides a natural way to
implement so-called "middleware". For instance, one can provide a
handler that simply writes all requests to a log file.

Handlers that wish to modify, but not respond to requests can use the
UpdateRequest method.