Remove unnecessary use of atomic package
The `handled` method of a `Twalk` request is always called from the
same goroutine that affects its return value (by calling `Rwalk`). Therefore,
there is no race condition, because `walker.filled` is never accessed by
another goroutine.

This is only true when combined with the other change in this commit that
moves `w.filled[el.index] = 1` into the `Rwalk` function and out of the walker's
`run` goroutine.