Simplify detection of unhandled requests and address race condition
Previously we used the `Session.conn.pendingReq` map to check if
a request had been handled by checking if there was a pending message
with its tag. This was incorrect, because in the time between when a
Handler clears a tag from this map, and when the server checks this map
for the presence of the tag, a client can send a new request with the
same tag. This is exactly what happened with highly parallel clients like
v9fs.

With this change, we make the observation that, in the case of a simple
Handler (single goroutine), the Handler's goroutine is the only one that
both answers requests (via the `R*` methods) and checks if they have
been answered. Therefore, it is safe for that goroutine to update a variable
in its `Session` struct (which is also private to each Handler).

Extending this observation to more complex Handlers, such as those
created by the `Stack` function, notice that each individual Handler
gets its own shallow copy of a `*Session`, so it is updating its own private
flag. This gives us the desired behavior; a Handler only cares if it answered
a request, and should not worry about downstream handlers. One small
change to the management handler allows us to propogate this status
up to the top.