Fix bug in styx.Stack with loop variables
This mistake took embarrassingly long for me to find out,
even though I was aware of the behavior. You cannot
reference a loop index variable in a goroutine closure and
expect it to be "frozen" at the current iteration; you will end
up getting the value of the variable at the time your goroutine
accesses it (usually ends up being the last value in the range).