Streamline API for responding to requests
Most response methods now take an error argument, and can directly
accept the return values of several of the `os` package functions. For
instance,

	fi, err := os.Stat(t.Path())
	if err != nil {
		t.Rerror("%s", err)
	} else {
		t.Rwalk(fi.Mode())
	}

becomes

	t.Rwalk(os.Stat(t.Path())

While this clearly shows some bias towards the `os` package, it is not
unreasonable to think that many other packages that work with file-like
objects will mimic the signatures of their analagous `os` package
functions.