Reduce the number of irrecoverable errors
With this change, the only errors that would end the session and
close the connection are those that result in an 'unspecified' state,
such as when a client uses a tag that is already in use for a pending
request.

In the future, this can allow for clients that reduce round trips by being
optimistic. That is, several requests can be pipelined together in a single
(network) packet if the client assumes that they will all succeed. This
added resilience also stems from my experience using a couple file
servers built with this package; it is much more jarring to lose the connection
when you've mounted the server to a directory, when compared to something
like refreshing a browser tab.