Do not respond to cancelled requests
If a request is cancelled by a flush transaction *before* its response is
received, the client should not expect to receive that response. This is
implied by the following language in flush(5):

          When the client sends a Tflush, it must wait to receive the
          corresponding Rflush before reusing oldtag for subsequent
          messages.

In other words, a client can re-use oldtag for future requests after
receiving the `Rflush` response. Therefore, responding to a flushed
request would be illegal, as the server would be responding to a tag
that is not in use, or in use for another request.