Close connection if tag is used twice
There is no obviously correct action for a server to take when a client
uses the same tag twice; responding with an Rerror message would
effectively cancel the previous request, not responding at all would
leave the client hanging. It only makes sense to handle this error by
closing the connection.